+++ lib/jotform.dart
@@ -1,1 +1,7 @@
 library jotform;
+
+export 'api.dart';
+export 'models/answer.dart';
+export 'models/answer_type.dart';
+export 'models/submission.dart';
+export 'models/submission_status.dart';