+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+* Added Submission & Answer models and generally cleaned up the json
+* Added basic api with limit tracker
+
 ## 1.0.0
 
 * Initial Release